add optional args for autofocus and focusNode
Adds configuration to control when the search field focuses. This is particularly helpful if the search is hidden and revealed through an animation.

`autofocus` defaults to true, so there the default functionality does not change. `focusNode` defaults to `null`, which is fine. `TextField` inline documentation says, "If null, this widget will create its own [FocusNode]."